pyproject.toml
[project]
name = "confengine-to-youtube"
version = "0.1.0"
description = "ConfEngineのセッション情報でYouTube動画を更新"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "budoux>=0.6",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.1",
    "google-auth-oauthlib>=1.0",
    "markdownify>=0.11",
    "pydantic>=2.0",
    "returns>=0.23",
    "rich>=13.0",
    "ruamel.yaml>=0.18",
    "snakemd>=2.0",
    "wcwidth>=0.2",
]

[dependency-groups]
dev = [
    "add-trailing-comma",
    "google-api-python-client-stubs",
    "google-auth-stubs",
    "import-linter",
    "mypy",
    "pytest",
    "pytest-cov",
    "ruff",
    "taskipy",
    "types-httplib2",
]

[project.scripts]
confengine-to-youtube = "confengine_to_youtube.infrastructure.cli:main"

[tool.uv]
package = true

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # フォーマッターと競合するルール（ruffが警告）
    "COM812", # missing-trailing-comma
    # 自明なコードにdocstringを強制しない
    "D100",   # missing-docstring-in-public-module
    "D101",   # missing-docstring-in-public-class
    "D102",   # missing-docstring-in-public-method
    "D103",   # missing-docstring-in-public-function
    "D104",   # missing-docstring-in-public-package
    "D105",   # missing-docstring-in-magic-method
    "D107",   # missing-docstring-in-__init__
    # 相互に競合するルール（D211を採用）
    "D203",   # incorrect-blank-line-before-class (D211と競合)
    # 相互に競合するルール（D212を採用）
    "D213",   # multi-line-summary-second-line (D212と競合)
    # 日本語docstringには句点不要
    "D400",   # first-line-should-end-with-period
    "D415",   # first-line-should-end-with-period-question-mark-or-exclamation-point
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004", # マジックナンバーの比較はテストで許容
    "S101",    # assert は pytest で必須
    "SLF001",  # プライベートメソッドのテストを許容
    "TC001",   # TYPE_CHECKING はテストでは不要
    "TC003",   # TYPE_CHECKING はテストでは不要
]

[tool.mypy]
python_version = "3.14"
strict = true
packages = ["confengine_to_youtube", "tests"]
plugins = ["returns.contrib.mypy.returns_plugin"]

[[tool.mypy.overrides]]
module = [
    "budoux",
    "google_auth_oauthlib.*",
    "markdownify",
    "ruamel.*",
    "snakemd.*",
    "wcwidth",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=confengine_to_youtube --cov-branch --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["src/confengine_to_youtube"]
omit = [
    "src/confengine_to_youtube/infrastructure/*",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.taskipy.tasks]
check = "task lint && task test"
format = "find src tests -name '*.py' -type f -exec add-trailing-comma {} + && ruff format"
lint = "ruff check && ruff format --check && mypy && lint-imports"
test = "pytest"

[tool.importlinter]
root_package = "confengine_to_youtube"
include_external_packages = true
contract_types = [
    "forbidden_external_submodule: confengine_to_youtube.import_contracts.ForbiddenExternalSubmoduleContract",
]

[[tool.importlinter.contracts]]
name = "Clean Architecture layers"
type = "layers"
layers = [
    "confengine_to_youtube.infrastructure",
    "confengine_to_youtube.adapters",
    "confengine_to_youtube.usecases",
    "confengine_to_youtube.domain",
]

[[tool.importlinter.contracts]]
name = "unsafe_perform_io is only allowed in infrastructure"
type = "forbidden_external_submodule"
source_modules = [
    "confengine_to_youtube.domain",
    "confengine_to_youtube.usecases",
    "confengine_to_youtube.adapters",
]
forbidden_external_modules = [
    "returns.unsafe",
]
